+++ single_server/src/server/validating.rs
@@ -23,7 +23,7 @@ impl MyLocationStorage {
         }
     }
 
-    fn parse_valid_idx(&self, idx : u32) -> Result<usize, Status> {
+    fn parse_valid_idx(&self, idx : u64) -> Result<usize, Status> {
         let res_idx = usize::try_from(idx);
         if res_idx.is_err() /*|| !self.timeline.is_point(idx.unwrap())*/ {
             return Err(Status::invalid_argument(format!("Not a valid id: {:}.", idx)));
@@ -31,7 +31,7 @@ impl MyLocationStorage {
         Ok(res_idx.unwrap())
     }
 
-    fn parse_valid_epoch(&self, epoch : u32) -> Result<usize, Status> {
+    fn parse_valid_epoch(&self, epoch : u64) -> Result<usize, Status> {
         let res_epoch = usize::try_from(epoch);
         if res_epoch.is_err() /*|| self.timeline.epochs() <= result_req_epoch.unwrap()*/ {
             return Err(Status::invalid_argument(format!("Not a valid epoch: {:}.", epoch)));
@@ -39,7 +39,7 @@ impl MyLocationStorage {
         Ok(res_epoch.unwrap())
     }
 
-    fn parse_valid_location(&self, x : u32, y : u32) -> Result<(usize, usize), Status> {
+    fn parse_valid_location(&self, x : u64, y : u64) -> Result<(usize, usize), Status> {
         let res_x = usize::try_from(x);
         let res_y = usize::try_from(y);
         if res_x.is_err() || res_y.is_err() /*|| self.timeline.xs() <= result_req_x.unwrap()*/ {
@@ -102,7 +102,7 @@ impl LocationStorage for MyLocationStorage {
                 (Err(err), _) | (_, Err(err)) => return Err(err),
         };
         match self.storage.get_user_location_at_epoch(epoch, req_idx) {
-            Some((x,y )) => Ok(Response::new(ObtainLocationReportResponse { pos_x : x as u32, pos_y : y as u32,})),
+            Some((x,y )) => Ok(Response::new(ObtainLocationReportResponse { pos_x : x as u64, pos_y : y as u64,})),
             None => Err(Status::not_found(format!("User with id {:} not found at epoch {:}", req_idx, epoch))),
         } 
     }
